permanently completed portfolio backend

+++ BackEnd/app/Http/Controllers/ProfileControllers/PortfolioController.php
@@ -4,12 +4,10 @@ namespace App\Http\Controllers\ProfileControllers;
 
 use App\Http\Controllers\Controller;
 use App\Models\Portfolio;
-use App\Models\PortfolioSkills;
 use App\Models\PortfolioFile;
 use Illuminate\Support\Arr;
 use Illuminate\Http\Request;
-use App\Http\Requests\AddPortfolioRequest;
-use App\Http\Requests\EditPortfolioRequest;
+use App\Http\Requests\PortfolioRequest;
 use App\Http\Resources\PortfolioResource;
 use App\Http\Resources\PortfolioCollection;
 
@@ -19,6 +17,8 @@ class PortfolioController extends Controller
     {
         // Get User
         $user = auth()->user();
+
+        // Check user
         if ($user == null) {
             return response()->json([
                 'user' => 'Invalid user'
@@ -41,7 +41,7 @@ class PortfolioController extends Controller
         ], 200);
     }
 
-    public function AddPortfolio(AddPortfolioRequest $request)
+    public function AddPortfolio(PortfolioRequest $request)
     {
         // Validate request
         $validated = $request->validated();
@@ -89,14 +89,32 @@ class PortfolioController extends Controller
         // Response
         return response()->json([
             "message" => "Portfolio created sucessfully",
-            "data" => new PortfolioCollection($portfolio)
+            "data" => new PortfolioResource($portfolio)
         ], 201);
     }
 
-    public function EditPortfolio(EditPortfolioRequest $request, Portfolio $portfolio)
+    public function EditPortfolio(PortfolioRequest $request, Portfolio $portfolio)
     {
         // Validate request
         $validated = $request->validated();
+
+        // Get User
+        $user = auth()->user();
+
+        // Check user
+        if ($user == null) {
+            return response()->json([
+                'user' => 'Invalid user'
+            ], 401);
+        }
+
+        // Check if portfolio belongs to user
+        if($user->id != $portfolio->user_id) {
+            return response()->json([
+                'user' => 'Invalid user'
+            ], 401);
+        }
+
         // Handle photo file
         if ($request->hasFile('photo')) {
             $Path = $request->file('photo')->store('avatars', 'public');
@@ -104,8 +122,7 @@ class PortfolioController extends Controller
         }
 
         $data = $validated;
-        
-       // $data = Arr::except($data, 'files');
+        $data = Arr::except($data, 'files');
         $data = Arr::except($data, 'skills');
         $portfolio->update($data);
 
@@ -119,12 +136,15 @@ class PortfolioController extends Controller
         // Remove the skills that are not in the validated skills
         $portfolio->skills()->whereIn('skill_id', $skillsToRemove)->delete();
 
+        // Add the new skills
         foreach($skills as $skill) {
             if(!in_array($skill, $currentSkills)) {
                 $portfolio->skills()->attach($skill);
-                $portfolio->save();
             }
         }
+
+        // Save skills
+        $portfolio->save();
 
         // Handle file uploads if present in the request
         if ($request->hasFile('files')) {
@@ -147,14 +167,33 @@ class PortfolioController extends Controller
         // Response
         return response()->json([
             "message" => "Portfolio updated sucessfully",
-            "data" => new PortfolioResource($portfolio),
+            "data" => new PortfolioResource($portfolio)
         ], 200);
     }
 
     public function DeletePortfolio(Request $request, Portfolio $portfolio)
     {
+        // Get User
+        $user = auth()->user();
+
+        // Check user
+        if ($user == null) {
+            return response()->json([
+                'user' => 'Invalid user'
+            ], 401);
+        }
+
+        // Check if portfolio belongs to user
+        if($user->id != $portfolio->user_id) {
+            return response()->json([
+                'user' => 'Invalid user'
+            ], 401);
+        }
+
+        // Delete portfolio
         $portfolio->delete();
 
+        // Response
         return response()->json([
             "message" => "Portfolio deleted",
         ], 202);
